[d3d8] Cast null pixel shader to DWORD
Stops GCC from whining.

+++ src/d3d8/d3d8_device.cpp
@@ -471,7 +471,7 @@ namespace dxvk {
     }
 
     if (Handle == DWORD(NULL)) {
-      m_currentPixelShader = NULL;
+      m_currentPixelShader = DWORD(NULL);
       return GetD3D9()->SetPixelShader(nullptr);
     }
